CMSC 140, Project 2
This program is a Movie Ticket Booking System implemented in C++. It allows users to select a movie, verify their age for suitability, and calculate the total cost of tickets for adults and children. It will also show the error messages for different variables, for example, if an invalid age is entered it will tell the user that this movie isn't suitable for the age the user entered.