CMSC 140, Project 3
This program is designed to simulate a hotel management system for the BlueMont Hotel chain. It collects data about the hotel's location, floors, and room occupancy, then calculates important metrics such as hotel income, total occupied and unoccupied rooms, and the occupancy rate. Additionally, it identifies the floor with the minimum number of rooms. The program also provides feedback for improving the hotel's occupancy rate.