CMSC 140, Project 4
This program is designed to track the absenteeism of employees in a company, calculate the total number of absent days, and determine the average number of days absent per employee. The program then saves this information into a file named employeeAbsences.txt. It also includes input validation to ensure data accuracy and provides a summary of the results.